chmod sock file dance
in order for fn to work without root, we must set permissions on the sock file
we create to allow users to write to the sock file. since go creates the file
in bind, we have to bind to a not-the-real sock file and hard link that file
to the real sock file, which will send an inotify event to fn to start sending
in tasks. voila. soft link didn't play nice.