umask allow users to write files
this seems like a bad idea. it is, however, a reasonably good idea since using
chmod will result in a race between creating the socket and setting the permissions
where fn could try to connect to it. with this, fn doesn't need to be run as
root, which we kinda need to satisfy really, and it seems a bit odd to try to
maneuver based on context of in a service or oss from the fdk level.

reset gopkg lock file (won't work until master is updated)