WIP: first rev of fdk 2.0
would like some eyes before going whole hog please :)

largely, this changes the context object to an interface, which makes doing
shenanigans with the headers nice and transparent to users, where they'll
access the headers the same whether they are http gateway headers or invoke
headers. this also drops support for all non-http-stream formats, be gone
pests. this also gets rid of the utils package and makes this all top level
again instead of coercing contexts back and forth, since we don't need to futz
with pipes for tests maybe we can get creative without having a separate
package for that stuff anymore.

useful points of feedback I'm asking for without making any presuppositions by
committing something (but don't let that starve your critical creativity):

do you want `hctx := fdk.GetHTTPContext(ctx)` ? should it return nil if it's
not an http context? should we make the user cast so that they have to panic
and not us? do lions have the ability to communicate with dolphins? one thing
to consider is that there is overlap between the http context and vanilla
context, so using the vanilla context type from `fdk.GetContext(ctx)` will
still allow getting at the (modified) headers and config, and getting at the
http context is only required for the request method and url (for users).

this totally breaks the API since now these things are addressed by interface
methods instead of fields on a concrete type. I should have been smacked 12
months ago, I blame everyone else no way this was my fault, nope. is this OK?
should we worry about compat? the changes are really easy in a lot of cases
(add parens...) and the trade offs seem ok (and we don't have a stable API and
the heat death of the universe is imminent?)

i really did try to make the godoc look nice. if you are anywhere near as anal
as i am i would appreciate a review of `$ godoc .`

i'm hesitant to make the out writer on the handler an http.ResponseWriter, but
it would make a whole lot of sense. kind of the whole point was to get away
from that, but in any event if we do want this, it's a good chance to do it.
we can change the handler signature if we'd like but that may be a slightly
harder breakage for users to fix up. open to ideas anyway.

... if this is the right path forward, I will add more tests to this and go
fix up the test image in fn (possibly punting the latter for a couple weeks
off when i'll have time to properly kill format in fn).