add http.Handler support for easiness
turns an http.Handler into an fdk.Handler... hopefully this is useful. the
error handling story there is... not great, feel free to send hate mail. but
it's pretty slick to unwind the fdk context at least... some of the stuff
makes me a little uneasy but it will probably all be okay.