change example to dockerfile
i'm doing this here so that somebody else can pick up and move this to
wherever it's supposed to go with properly formatted documentation, this was
the closest working example that I could find of a go program laying around to
use to exhibit this (fn's examples are broken, and the tutorials repo is only
README, not deployable code). this directory itself is slated for removal, to
be centralized in some unknown place with the other lang tutorials and this
repo has a simple godoc friendly example at the top level already, to go along
with the normal 'go intro' in the tutorials directory.