Ensure release script actually bumps version.go
The release script needs to push the version bump change it makes to version.go - master is currently failing because we're trying to push tags for a version that already exists. This also bumps current contents of version.go so that is in line with the tags/releases that already exist.